pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/on_this_day"]

[project]
name = "on-this-day-trmnl"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "Pillow>=10.0",
    "python-dotenv>=1.0",
    "tzdata>=2024.1",
]

[project.scripts]
on-this-day = "on_this_day.__main__:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
]
